Move tile array/map lowering logic outside JAX bindings.
This PR is moving the Poplar lowering logic into proper functions, in order to make the JAX/XLA custom ops layer as thin as possible.
It should allow us to support additional backends (e.g. Poptorch) in the near future.
