+++ setup.cfg
@@ -9,30 +9,3 @@ per-file-ignores =
     tessellate_ipu/lax/__init__.py: F401
     tessellate_ipu/linalg/__init__.py: F401
     tessellate_ipu/utils/__init__.py: F401
-
-[isort]
-line_length = 120
-known_first_party = tessellate_ipu
-# multi_line_output = 3
-# include_trailing_comma = True
-# force_grid_wrap = 0
-# combine_as_imports = True
-
-[mypy]
-plugins = numpy.typing.mypy_plugin
-# Config heavily inspired by Pydantic!
-python_version = 3.8
-show_error_codes = True
-# strict_optional = True
-warn_redundant_casts = True
-warn_unused_ignores = True
-warn_unused_configs = True
-check_untyped_defs = True
-disallow_any_generics = True
-no_implicit_optional = False
-# disallow_incomplete_defs = True
-# disallow_untyped_decorators = True
-# disallow_untyped_calls = True
-# # disallow_subclassing_any = True
-# # for strict mypy: (this is the tricky one :-))
-# disallow_untyped_defs = True
